Fix potential overflow in compression code. It is not clear to me whether this overflow has actually occurred (the more threads, the less likely). If it did, it may have had an impact on compression ratio. (Changing "avg" from uint32_t to uint64_t shouldn't change anything.)

+++ src/compress.c
@@ -79,8 +79,8 @@ struct {
 } paircands[MAX_NEW];
 
 static uint8_t newtest[MAXSYMB][MAXSYMB];
-static uint32_t (*countfirst)[MAX_NEW][MAXSYMB];
-static uint32_t (*countsecond)[MAX_NEW][MAXSYMB];
+static uint64_t (*countfirst)[MAX_NEW][MAXSYMB];
+static uint64_t (*countsecond)[MAX_NEW][MAXSYMB];
 
 static struct Work *work = NULL, *work_adj = NULL;
